requirements.txt
crewai==0.126.0
gradio>=4.44.1
pydantic>=2.0.0
fastapi>=0.100.0
uvicorn>=0.23.0
sqlalchemy>=2.0.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
numpy>=1.24.0
opencv-python>=4.8.0
gTTS>=2.3.2
moviepy>=1.0.3
torch>=2.0.0
transformers>=4.30.0
langchain>=0.0.267
